tests: combine library permissions tests
* test_library_permissions.py
* test_library_item_permissions.py

becomes just test_libraries_permissions.py
